+++ integreat_cms/api/v3/locations.py
@@ -91,7 +91,8 @@ def transform_poi_translation(poi_translation):
         "location": transform_poi(poi),
         "hash": None,
         "stops": poi.stops or None,
-        "spoken_languages": poi.spoken_languages or None, 
+        "spoken_languages": poi.spoken_languages or None,
+        "target_group": poi.target_group or None,
     }